(rescue agent) Add the Rescue agent.
When the agent starts it creates a rescue agent with the copy of the
working agent.

When the agent does not starts or if it enters into a loop, we reinstall
the agent with the "rescue image".

This allow to fix broken updates, etc.

We creates a BOOL_FILE_CONTROL_WATCH_DOG watchdog file, to ensure that the
pulse agent runs smoothly