(deploy) Remove support of the pulse user
When we started pulse, the user used to deploy the packges on the
clients was 'pulse'.
Later we worked on the linux agent. But the pulse user was already used
by pulseaudio.
Because of this we changed and used pulseuser. We kept on windows the
compatibility with the pulse user.

Now that this is an old story and everybody is using recent enough pulse
we can remove the use of this user.